+++ client/src/app/components/search/search.module.ts
@@ -2,6 +2,7 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MatButtonModule } from '@angular/material/button';
+import { MatDialogModule } from '@angular/material/dialog';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatIconModule } from '@angular/material/icon';
 import { MatInputModule } from '@angular/material/input';
@@ -13,14 +14,16 @@ import { TimeagoModule } from 'ngx-timeago';
 import { InputComponent } from './input/input.component';
 import { PostComponent } from './post/post.component';
 import { WrapperComponent } from './wrapper/wrapper.component';
+import { NotificationDialogComponent } from './notification-dialog/notification-dialog.component';
 
 @NgModule({
-  declarations: [WrapperComponent, InputComponent, PostComponent],
+  declarations: [WrapperComponent, InputComponent, PostComponent, NotificationDialogComponent],
   imports: [
     CommonModule,
     FormsModule,
     ReactiveFormsModule,
     MatButtonModule,
+    MatDialogModule,
     MatFormFieldModule,
     MatInputModule,
     MatIconModule,
@@ -29,6 +32,9 @@ import { WrapperComponent } from './wrapper/wrapper.component';
     MatSnackBarModule,
     MatSlideToggleModule,
     TimeagoModule.forChild()
-  ]
+  ], 
+  entryComponents: [
+    NotificationDialogComponent
+  ],
 })
 export class SearchModule { }